+++ packages/guide/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added animals to ranch shop
 - Added support for korean and brazilian portuguese
+- Added separator to locations and sorted shops alphabetically
 
 ### 0.13.0 2023-11-01